+++ src/json/NLohmannJsonDarkSky.cpp
@@ -19,8 +19,8 @@
 */
 
 #include "NLohmannJsonDarkSky.hpp"
-#include <cmath>
 #include <iostream>
+#include "../util/NumericPrecision.hpp"
 
 namespace wic
 {
@@ -126,8 +126,8 @@ bool NLohmannJsonDarkSky::parseSingleWeatherItem(const value_type& dataPoint, We
     // calculate them on the fly.
     weather.setTemperatureFahrenheit(weather.temperatureCelsius() * 1.8 + 32);
     // Avoid values like 6.9999... ° F by rounding, if appropriate.
-    const float fahrenheitRounded = std::round(weather.temperatureFahrenheit());
-    if (std::fabs(fahrenheitRounded - weather.temperatureFahrenheit()) < 0.005)
+    const float fahrenheitRounded = NumericPrecision<float>::enforce(weather.temperatureFahrenheit());
+    if (fahrenheitRounded != weather.temperatureFahrenheit())
     {
       weather.setTemperatureFahrenheit(fahrenheitRounded);
     }