+++ src/json/NLohmannJsonWeatherstack.cpp
@@ -19,13 +19,13 @@
 */
 
 #include "NLohmannJsonWeatherstack.hpp"
-#include <cmath>
 #include <ctime>
 #include <iomanip>
 #include <iostream>
 #include <sstream>
 #include "../../third-party/nlohmann/json.hpp"
 #include "WeatherstackFunctions.hpp"
+#include "../util/NumericPrecision.hpp"
 
 namespace wic
 {
@@ -102,8 +102,8 @@ bool NLohmannJsonWeatherstack::parseCurrentWeather(const std::string& json, Weat
       // calculate them on the fly.
       weather.setTemperatureFahrenheit(static_cast<float>(weather.temperatureCelsius() * 1.8 + 32.0));
       // Avoid values like 6.9999... ° F by rounding, if appropriate.
-      const float fahrenheitRounded = std::round(weather.temperatureFahrenheit());
-      if (std::fabs(fahrenheitRounded - weather.temperatureFahrenheit()) < 0.005)
+      const float fahrenheitRounded = NumericPrecision<float>::enforce(weather.temperatureFahrenheit());
+      if (fahrenheitRounded != weather.temperatureFahrenheit())
       {
         weather.setTemperatureFahrenheit(fahrenheitRounded);
       }