+++ changelog.md
@@ -1,5 +1,11 @@
 # Version history of weather-information-collector
 
+## Version 0.9.17-pre (2021-04-25)
+
+Precision of temperature values is now limited to three decimal digits after
+the decimal separator, i. e. values like 23.4560001 °C will become 23.456 °C
+instead.
+
 ## Version 0.9.16 (2021-04-23)
 
 - Parser errors will now show more information why JSON parsing failed.